added custom Error in all resp & added JSdocs

+++ src/controllers/user.js
@@ -1,7 +1,15 @@
 const User = require("../models/user");
 const { matchedData } = require("express-validator");
 const { userResponse } = require("../utils/helper");
+const { HttpError } = require("../utils/helper");
 
+/**
+ * @description Get user data by `userId`
+ * @param req {object} Express req object
+ * @param res {object} Express res object
+ * @param next {function} Express next middleware callback
+ * @returns {Promise<*>}
+ */
 const getUserById = async (req, res, next) => {
   try {
     const { id } = req.params;
@@ -9,7 +17,7 @@ const getUserById = async (req, res, next) => {
     const user = await User.findById(id).exec();
 
     if (!user) {
-      throw new Error("User not found");
+      throw new HttpError(404, "User not found");
     }
 
     return res.json({
@@ -22,6 +30,13 @@ const getUserById = async (req, res, next) => {
   }
 };
 
+/**
+ * @description Get user data of a Signed-in user
+ * @param req {object} Express req object
+ * @param res {object} Express res object
+ * @param next {function} Express next middleware callback
+ * @returns {Promise<*>}
+ */
 const getUser = async (req, res, next) => {
   try {
     const { id } = req.auth;
@@ -29,7 +44,7 @@ const getUser = async (req, res, next) => {
     const user = await User.findById(id).exec();
 
     if (!user) {
-      throw new Error("User not found");
+      throw new HttpError(404, "User not found");
     }
 
     return res.json({
@@ -42,15 +57,24 @@ const getUser = async (req, res, next) => {
   }
 };
 
+/**
+ * @description Get all users data (Admin) *TODO: paginate & sort*
+ * @param req {object} Express req object
+ * @param res {object} Express res object
+ * @param next {function} Express next middleware callback
+ * @returns {Promise<*>}
+ */
 const getAllUsers = async (req, res, next) => {
   try {
-    const { type } = req.query;
-    const role = type === "customer" || "seller" || "admin" ? type : undefined;
+    // const { type } = req.query;
+    // const role = type === "customer" || "seller" || "admin" ? type : undefined;
 
-    const users = await User.find({ /*role*/ }).exec();
+    const users = await User.find({
+      /*role*/
+    }).exec();
 
     if (users.length < 1) {
-      throw new Error("No User found");
+      throw new HttpError(404, "No User found");
     }
 
     return res.json({
@@ -63,6 +87,13 @@ const getAllUsers = async (req, res, next) => {
   }
 };
 
+/**
+ * @description Update user data of a Signed-in user
+ * @param req {object} Express req object
+ * @param res {object} Express res object
+ * @param next {function} Express next middleware callback
+ * @returns {Promise<*>}
+ */
 const updateUser = async (req, res, next) => {
   try {
     const { id } = req.auth;
@@ -84,6 +115,13 @@ const updateUser = async (req, res, next) => {
   }
 };
 
+/**
+ * @description Update user data by `userId` (Admin)
+ * @param req {object} Express req object
+ * @param res {object} Express res object
+ * @param next {function} Express next middleware callback
+ * @returns {Promise<*>}
+ */
 const updateUserById = async (req, res, next) => {
   try {
     const { id } = req.params;
@@ -106,14 +144,25 @@ const updateUserById = async (req, res, next) => {
   }
 };
 
+/**
+ * @description Delete user by `userId` (Admin)
+ * @param req {object} Express req object
+ * @param res {object} Express res object
+ * @param next {function} Express next middleware callback
+ * @returns {Promise<*>}
+ */
 const deleteUserById = async (req, res, next) => {
   try {
     const { id } = req.params;
 
+    if (req.auth.id === id) {
+      throw new HttpError(403, "User cannot delete himself");
+    }
+
     const { deletedCount } = await User.deleteOne({ _id: id }).exec();
 
     if (!deletedCount) {
-      throw Error("no user found to delete");
+      throw new HttpError(404, "no user found to delete");
     }
 
     return res.json({
